Fix: Abp.Dapper test failing on appveyor #2201 Enhancement: Execute & ExecuteAsync should return int & Task<int> #2290
