+++ src/Abp.Dapper/Dapper/Repositories/IDapperRepositoryOfTEntityAndTPrimaryKey.cs
@@ -229,16 +229,17 @@ namespace Abp.Dapper.Repositories
         /// <summary>
         ///     Executes the given query text
         /// </summary>
-        /// <param name="query"></param>
-        /// <param name="parameters"></param>
-        void Execute([NotNull] string query, [CanBeNull] object parameters = null);
+        /// <param name="query">The query.</param>
+        /// <param name="parameters">The parameters.</param>
+        /// <returns></returns>
+        int Execute([NotNull] string query, [CanBeNull] object parameters = null);
 
         /// <summary>
         ///     Executes as async the given query text
         /// </summary>
         /// <param name="query"></param>
         /// <param name="parameters"></param>
-        Task ExecuteAsync([NotNull] string query, [CanBeNull] object parameters = null);
+        Task<int> ExecuteAsync([NotNull] string query, [CanBeNull] object parameters = null);
 
         /// <summary>
         ///     Gets the set.
