+++ test/Abp.Dapper.Tests/AbpDapperTestModule.cs
@@ -1,13 +1,12 @@
 ﻿using System.Collections.Generic;
-using System.Data.SqlClient;
-using System.IO;
 using System.Reflection;
+using System.Transactions;
 
 using Abp.EntityFramework;
 using Abp.Modules;
 using Abp.TestBase;
 
-using Dapper;
+using DapperExtensions.Sql;
 
 namespace Abp.Dapper.Tests
 {
@@ -18,43 +17,18 @@ namespace Abp.Dapper.Tests
     )]
     public class AbpDapperTestModule : AbpModule
     {
+        public override void PreInitialize()
+        {
+            Configuration.UnitOfWork.IsolationLevel = IsolationLevel.Unspecified;
+        }
+
         public override void Initialize()
         {
             IocManager.RegisterAssemblyByConvention(Assembly.GetExecutingAssembly());
 
+            DapperExtensions.DapperExtensions.SqlDialect = new SqliteDialect();
+
             DapperExtensions.DapperExtensions.SetMappingAssemblies(new List<Assembly> { Assembly.GetExecutingAssembly() });
-        }
-
-        public override void Shutdown()
-        {
-            var connection = new SqlConnection(Configuration.DefaultNameOrConnectionString);
-
-            var files = new List<string>
-            {
-                ReadScriptFile("DestroyScript")
-            };
-
-            foreach (string setupFile in files)
-            {
-                connection.Execute(setupFile);
-            }
-        }
-
-        private string ReadScriptFile(string name)
-        {
-            string fileName = GetType().Namespace + ".Scripts" + "." + name + ".sql";
-            using (Stream resource = Assembly.GetExecutingAssembly().GetManifestResourceStream(fileName))
-            {
-                if (resource != null)
-                {
-                    using (var sr = new StreamReader(resource))
-                    {
-                        return sr.ReadToEnd();
-                    }
-                }
-            }
-
-            return string.Empty;
         }
     }
 }
