--- test/Abp.Dapper.Tests/SampleDapperApplicationDbContext.cs
@@ -1,36 +1,0 @@
-﻿using System.Data.Common;
-using System.Data.Entity;
-using System.Data.Entity.SqlServer;
-
-using Abp.Dapper.Tests.Entities;
-using Abp.EntityFramework;
-
-namespace Abp.Dapper.Tests
-{
-    [DbConfigurationType(typeof(DapperDbContextConfiguration))]
-    public class SampleDapperApplicationDbContext : AbpDbContext
-    {
-        public SampleDapperApplicationDbContext(DbConnection connection)
-            : base(connection, false)
-        {
-        }
-
-        public SampleDapperApplicationDbContext(DbConnection connection, bool contextOwnsConnection)
-            : base(connection, contextOwnsConnection)
-        {
-        }
-
-        public virtual IDbSet<Product> Products { get; set; }
-    }
-
-    public class DapperDbContextConfiguration : DbConfiguration
-    {
-        public DapperDbContextConfiguration()
-        {
-            SetProviderServices(
-                "System.Data.SqlClient",
-                SqlProviderServices.Instance
-            );
-        }
-    }
-}
